requirements.txt
requests
supabase
openai
python-dotenv